+++ triplea/cli/aaa.py
@@ -22,5 +22,8 @@ click.echo(f"Title: {tr.__title__} ({tr.__version__})")
 click.echo(f"   {tr.__description__}")
 click.echo()
 
+def main():
+    cli()
+
 if __name__ == "__main__":
     cli()