+++ triplea/config/settings.py
@@ -7,15 +7,22 @@ import tomli
 
 # Project Directories
 ROOT = pathlib.Path(__file__).resolve().parent.parent
-DB_ROOT_PATH = ROOT.parent / "database"
-ENV_PATH_FILE = ROOT / "config" / "environment_variable" / ".env"
+# DB_ROOT_PATH = ROOT.parent / "database"
+DB_ROOT_PATH = os.path.join( os.path.abspath(os.curdir)  , "database") 
+
+# ENV_PATH_FILE = ROOT / "config" / "environment_variable" / ".env"
+ENV_PATH_FILE = os.path.join( os.path.abspath(os.curdir)  , ".env") # For handelling in package
 
 load_dotenv(ENV_PATH_FILE, override=True)
 
-with open("pyproject.toml", "rb") as f:
-    pyproject = tomli.load(f)
-    version = pyproject["tool"]["poetry"]["version"]
+# with open("pyproject.toml", "rb") as f:
+#     pyproject = tomli.load(f)
+#     version = pyproject["tool"]["poetry"]["version"]
 
+# # https://stackoverflow.com/questions/67085041/how-to-specify-version-in-only-one-place-when-using-pyproject-toml
+import importlib.metadata
+# version = importlib.metadata.version(__package__ or __name__)
+version = importlib.metadata.version('triplea')
 
 class Settings(BaseSettings):
     # ---------------My Envirement Varable-------------------------------
@@ -44,7 +51,7 @@ class Settings(BaseSettings):
     )
 
     VERSION: Optional[str] = (
-        version + ".001"
+        version + ".002"
     )  # Change this micro version in the development process
 
     # class Config: